+++ src/polymer-project.ts
@@ -18,13 +18,12 @@ import * as osPath from 'path';
 import * as logging from 'plylog';
 import {Transform} from 'stream';
 import File = require('vinyl');
-import * as vfs from 'vinyl-fs';
+import {src as vinylSrc} from 'vinyl-fs';
 import {ProjectConfig, ProjectOptions} from 'polymer-project-config';
 
-import {StreamAnalyzer} from './analyzer';
+import {BuildAnalyzer} from './analyzer';
 import {Bundler} from './bundle';
 import {FileCB} from './streams';
-import {forkStream} from './fork-stream';
 
 const logger = logging.getLogger('polymer-project');
 const pred = dom5.predicates;
@@ -48,7 +47,7 @@ export class PolymerProject {
    * can be used to get information on dependencies and fragments for the
    * project once the source & dependency streams have been piped into it.
    */
-  analyzer: StreamAnalyzer;
+  analyzer: BuildAnalyzer;
 
   /**
    * A `Transform` stream that modifies the files that pass through it based
@@ -71,31 +70,29 @@ export class PolymerProject {
 
     logger.debug(`config: ${this.config}`);
 
-    this.analyzer = new StreamAnalyzer(this.config);
+    this.analyzer = new BuildAnalyzer(this.config);
     this.bundler = new Bundler(this.config, this.analyzer);
   }
 
-  // TODO(justinfagnani): add options, pass to vfs.src()
   /**
-   * Returns a streams of this project's source files - files matched by the
-   * globs returns from `getSourceGlobs`, and not matched by
-   * `getDependencyGlobs` (which are inverted and appended to the source globs).
+   * Returns the analyzer's stream of this project's source files - files
+   * matched by the project's `config.sources` value.
    */
   sources(): NodeJS.ReadableStream {
-    return vfs.src(this.config.sources, {
-      cwdbase: true,
-      nodir: true,
-    });
+    return this.analyzer.sources;
   }
 
+  /**
+   * Returns the analyzer's stream of this project's dependency files - files
+   * loaded inside the analyzed project that are not considered source files.
+   */
   dependencies(): NodeJS.ReadableStream {
-    let dependenciesStream: NodeJS.ReadableStream =
-        forkStream(this.analyzer.dependencies);
+    let dependenciesStream: NodeJS.ReadableStream = this.analyzer.dependencies;
 
-    // If we need to include additional dependencies, create a new vfs.src
+    // If we need to include additional dependencies, create a new vinyl source
     // stream and pipe our default dependencyStream through it to combine.
     if (this.config.extraDependencies.length > 0) {
-      const includeStream = vfs.src(this.config.extraDependencies, {
+      const includeStream = vinylSrc(this.config.extraDependencies, {
         cwdbase: true,
         nodir: true,
         passthrough: true,