+++ test/bundle_test.js
@@ -17,9 +17,7 @@
 const assert = require('chai').assert;
 const dom5 = require('dom5');
 const parse5 = require('parse5');
-const File = require('vinyl');
 const path = require('path');
-const stream = require('stream');
 const mergeStream = require('merge-stream');
 const ProjectConfig = require('polymer-project-config').ProjectConfig;
 
@@ -27,30 +25,23 @@ const analyzer = require('../lib/analyzer');
 const bundle = require('../lib/bundle');
 
 const Bundler = bundle.Bundler;
-const StreamAnalyzer = analyzer.StreamAnalyzer;
+const BuildAnalyzer = analyzer.BuildAnalyzer;
 
-const root = path.resolve('root/');
+const root = path.resolve('test/static/bundler-data');
 
 suite('Bundler', () => {
 
   let bundler;
-  let sourceStream;
   let bundledStream;
   let files;
 
   let setupTest = (options) => new Promise((resolve, reject) => {
-    options.root = root;
-    options.sources = options.files.map((f) => f.path);
-
+    options.root = options.root || root;
     let config = new ProjectConfig(options);
-    let analyzer = new StreamAnalyzer(config);
+    let analyzer = new BuildAnalyzer(config);
     bundler = new Bundler(config, analyzer);
-    sourceStream = new stream.Readable({
-      objectMode: true,
-    });
-    bundledStream = mergeStream(sourceStream, analyzer.dependencies)
-                        .pipe(analyzer)
-                        .pipe(bundler);
+    bundledStream =
+        mergeStream(analyzer.sources, analyzer.dependencies).pipe(bundler);
     files = new Map();
     bundledStream.on('data', (file) => {
       files.set(file.path, file);
@@ -61,20 +52,13 @@ suite('Bundler', () => {
     bundledStream.on('error', (err) => {
       reject(err);
     });
-    pushFiles(options.files);
   });
 
   teardown(() => {
     bundler = null;
-    sourceStream = null;
     bundledStream = null;
     files = null;
   });
-
-  const pushFiles = (files) => {
-    files.forEach((f) => sourceStream.push(f));
-    sourceStream.push(null);
-  };
 
   const getFile =
       (filename) => {
@@ -102,57 +86,61 @@ suite('Bundler', () => {
     return link != null;
   };
 
-  test('entrypoint only', () => setupTest({
-                                  entrypoint: 'entrypointOnly.html',
-                                  files: [framework(), entrypointOnly()],
-                                }).then((files) => {
-    const doc = parse5.parse(getFile('entrypointOnly.html'));
-    assert.isTrue(hasMarker(doc, 'framework'));
-    assert.isFalse(hasImport(doc, 'framework.html'));
-    // TODO(justinfagnani): check that shared-bundle.html doesn't exist
-    // it's in the analyzer's file map for some reason
-  }));
+  test(
+      'entrypoint only',
+      () => setupTest({
+              entrypoint: 'entrypoint-only.html',
+              sources: ['framework.html', 'entrypoint-only.html'],
+            }).then((files) => {
+        const doc = parse5.parse(getFile('entrypoint-only.html'));
+        assert.isTrue(hasMarker(doc, 'framework'));
+        assert.isFalse(hasImport(doc, 'framework.html'));
+        assert.isNotOk(getFile('shared-bundle.html'));
+      }));
 
-  test('two fragments', () => setupTest({
-                                fragments: ['shell.html', 'entrypointA.html'],
-                                files: [framework(), shell(), entrypointA()],
-                              }).then((files) => {
-    // shell doesn't import framework
-    const shellDoc = parse5.parse(getFile('shell.html'));
-    assert.isFalse(hasMarker(shellDoc, 'framework'));
-    assert.isFalse(hasImport(shellDoc, 'framework.html'));
+  test(
+      'two fragments',
+      () => setupTest({
+              entrypoint: 'entrypoint-a.html',
+              fragments: ['shell.html', 'entrypoint-a.html'],
+              sources: ['shell.html', 'entrypoint-a.html', 'framework.html'],
+            }).then((files) => {
+        // shell doesn't import framework
+        const shellDoc = parse5.parse(getFile('shell.html'));
+        assert.isFalse(hasMarker(shellDoc, 'framework'));
+        assert.isFalse(hasImport(shellDoc, 'framework.html'));
 
-    // entrypoint doesn't import framework
-    const entrypointDoc = parse5.parse(getFile('entrypointA.html'));
-    assert.isFalse(hasMarker(entrypointDoc, 'framework'));
-    assert.isFalse(hasImport(entrypointDoc, 'framework.html'));
+        // entrypoint doesn't import framework
+        const entrypointDoc = parse5.parse(getFile('entrypoint-a.html'));
+        assert.isFalse(hasMarker(entrypointDoc, 'framework'));
+        assert.isFalse(hasImport(entrypointDoc, 'framework.html'));
 
-    // No shared-bundle bundles framework
-    const sharedDoc = parse5.parse(getFile('shared-bundle.html'));
-    assert.isTrue(hasMarker(sharedDoc, 'framework'));
-    assert.isFalse(hasImport(sharedDoc, 'framework.html'));
+        // No shared-bundle bundles framework
+        const sharedDoc = parse5.parse(getFile('shared-bundle.html'));
+        assert.isTrue(hasMarker(sharedDoc, 'framework'));
+        assert.isFalse(hasImport(sharedDoc, 'framework.html'));
 
-    // fragments import shared-bundle
-    assert.isTrue(hasImport(entrypointDoc, 'shared-bundle.html'));
-    assert.isTrue(hasImport(shellDoc, 'shared-bundle.html'));
-  }));
+        // fragments import shared-bundle
+        assert.isTrue(hasImport(entrypointDoc, 'shared-bundle.html'));
+        assert.isTrue(hasImport(shellDoc, 'shared-bundle.html'));
+      }));
 
   test.skip(
       'shell and entrypoint',
       () => setupTest({
-              entrypoint: '/root/entrypointA.html',
-              shell: '/root/shell.html',
-              files: [framework(), shell(), entrypointA()],
+              entrypoint: 'entrypoint-a.html',
+              shell: 'shell.html',
+              files: ['framework.html', 'shell.html', 'entrypoint-a.html'],
             }).then((files) => {
+
         // shell bundles framework
         const shellDoc = parse5.parse(getFile('shell.html'));
         assert.isTrue(hasMarker(shellDoc, 'framework'));
-        assert.isFalse(hasImport(shellDoc, '/root/framework.html'));
 
         // entrypoint doesn't import framework
-        const entrypointDoc = parse5.parse(getFile('entrypointA.html'));
+        const entrypointDoc = parse5.parse(getFile('entrypoint-a.html'));
         assert.isFalse(hasMarker(entrypointDoc, 'framework'));
-        assert.isFalse(hasImport(entrypointDoc, '/root/framework.html'));
+        assert.isFalse(hasImport(entrypointDoc, 'framework.html'));
 
         // entrypoint imports shell
         assert.isTrue(hasImport(entrypointDoc, 'shell.html'));
@@ -164,14 +152,16 @@ suite('Bundler', () => {
   test(
       'shell and fragments with shared dependency',
       () => setupTest({
+              entrypoint: 'entrypoint-a.html',
               shell: 'shell.html',
-              fragments: ['entrypointB.html', 'entrypointC.html'],
-              files: [
-                framework(),
-                shell(),
-                entrypointB(),
-                entrypointC(),
-                commonDep(),
+              fragments: ['entrypoint-b.html', 'entrypoint-c.html'],
+              sources: [
+                'framework.html',
+                'shell.html',
+                'entrypoint-a.html',
+                'entrypoint-b.html',
+                'entrypoint-c.html',
+                'common-dependency.html',
               ],
             }).then((files) => {
         // shell bundles framework
@@ -181,17 +171,17 @@ suite('Bundler', () => {
 
         // shell bundles commonDep
         assert.isTrue(hasMarker(shellDoc, 'commonDep'));
-        assert.isFalse(hasImport(shellDoc, 'commonDep.html'));
+        assert.isFalse(hasImport(shellDoc, 'common-dependency.html'));
 
         // entrypoint B doesn't import commonDep
-        const entrypointBDoc = parse5.parse(getFile('entrypointB.html'));
+        const entrypointBDoc = parse5.parse(getFile('entrypoint-b.html'));
         assert.isFalse(hasMarker(entrypointBDoc, 'commonDep'));
-        assert.isFalse(hasImport(entrypointBDoc, 'commonDep.html'));
+        assert.isFalse(hasImport(entrypointBDoc, 'common-dependency.html'));
 
         // entrypoint C doesn't import commonDep
-        const entrypointCDoc = parse5.parse(getFile('entrypointC.html'));
+        const entrypointCDoc = parse5.parse(getFile('entrypoint-c.html'));
         assert.isFalse(hasMarker(entrypointCDoc, 'commonDep'));
-        assert.isFalse(hasImport(entrypointCDoc, 'commonDep.html'));
+        assert.isFalse(hasImport(entrypointCDoc, 'common-dependency.html'));
 
         // entrypoints import shell
         assert.isTrue(hasImport(entrypointBDoc, 'shell.html'));
@@ -201,100 +191,58 @@ suite('Bundler', () => {
         assert.isNotOk(getFile('shared-bundle.html'));
       }));
 
-  test.skip(
-      'entrypoint and fragments', () => setupTest({
-                                          entrypoint: '/root/entrypointA.html',
-                                          fragments: [
-                                            '/root/shell.html',
-                                            '/root/entrypointB.html',
-                                            '/root/entrypointC.html',
-                                          ],
-                                          files: [
-                                            framework(),
-                                            shell(),
-                                            entrypointA(),
-                                            entrypointB(),
-                                            entrypointC(),
-                                            commonDep(),
-                                          ],
-                                        }).then((files) => {
-        // shared bundle was emitted
-        const bundle = getFile('shared-bundle.html');
-        assert.ok(bundle);
-        const bundleDoc = parse5.parse(bundle);
+  test.skip('entrypoint and fragments', () => setupTest({
+                                                entrypoint: 'entrypoint-a.html',
+                                                fragments: [
+                                                  'shell.html',
+                                                  'entrypoint-b.html',
+                                                  'entrypoint-c.html',
+                                                ],
+                                                files: [
+                                                  'framework.html',
+                                                  'shell.html',
+                                                  'entrypoint-b.html',
+                                                  'entrypoint-c.html',
+                                                  'common-dependency.html',
+                                                ],
+                                              }).then((files) => {
+    // shared bundle was emitted
+    const bundle = getFile('shared-bundle.html');
+    assert.ok(bundle);
+    const bundleDoc = parse5.parse(bundle);
 
-        // shared-bundle bundles framework
-        assert.isTrue(hasMarker(bundleDoc, 'framework'));
-        assert.isFalse(hasImport(bundleDoc, '/root/framework.html'));
+    // shared-bundle bundles framework
+    assert.isTrue(hasMarker(bundleDoc, 'framework'));
+    assert.isFalse(hasImport(bundleDoc, 'framework.html'));
 
-        // shared-bundle bundles commonDep
-        assert.isTrue(hasMarker(bundleDoc, 'commonDep'));
-        assert.isFalse(hasImport(bundleDoc, '/root/commonDep.html'));
+    // shared-bundle bundles commonDep
+    assert.isTrue(hasMarker(bundleDoc, 'common-dependency'));
+    assert.isFalse(hasImport(bundleDoc, 'common-dependency.html'));
 
-        // entrypoint doesn't import framework
-        const entrypointDoc = parse5.parse(getFile('entrypointA.html'));
-        assert.isFalse(hasMarker(entrypointDoc, 'framework'));
-        assert.isFalse(hasImport(entrypointDoc, '/root/framework.html'));
+    // entrypoint doesn't import framework
+    const entrypointDoc = parse5.parse(getFile('entrypoint-a.html'));
+    assert.isFalse(hasMarker(entrypointDoc, 'framework'));
+    assert.isFalse(hasImport(entrypointDoc, 'framework.html'));
 
-        // shell doesn't import framework
-        const shellDoc = parse5.parse(getFile('entrypointA.html'));
-        assert.isFalse(hasMarker(shellDoc, 'framework'));
-        assert.isFalse(hasImport(shellDoc, '/root/framework.html'));
+    // shell doesn't import framework
+    const shellDoc = parse5.parse(getFile('entrypoint-a.html'));
+    assert.isFalse(hasMarker(shellDoc, 'framework'));
+    assert.isFalse(hasImport(shellDoc, 'framework.html'));
 
-        // entrypoint B doesn't import commonDep
-        const entrypointBDoc = parse5.parse(getFile('entrypointB.html'));
-        assert.isFalse(hasMarker(entrypointBDoc, 'commonDep'));
-        assert.isFalse(hasImport(entrypointBDoc, '/root/commonDep.html'));
+    // entrypoint B doesn't import commonDep
+    const entrypointBDoc = parse5.parse(getFile('entrypoint-b.html'));
+    assert.isFalse(hasMarker(entrypointBDoc, 'commonDep'));
+    assert.isFalse(hasImport(entrypointBDoc, 'common-dependency.html'));
 
-        // entrypoint C doesn't import commonDep
-        const entrypointCDoc = parse5.parse(getFile('entrypointC.html'));
-        assert.isFalse(hasMarker(entrypointCDoc, 'commonDep'));
-        assert.isFalse(hasImport(entrypointCDoc, '/root/commonDep.html'));
+    // entrypoint C doesn't import commonDep
+    const entrypointCDoc = parse5.parse(getFile('entrypoint-c.html'));
+    assert.isFalse(hasMarker(entrypointCDoc, 'commonDep'));
+    assert.isFalse(hasImport(entrypointCDoc, 'common-dependency.html'));
 
-        // entrypoint and fragments import shared-bundle
-        assert.isTrue(hasImport(entrypointDoc, 'shared-bundle.html'));
-        assert.isTrue(hasImport(entrypointBDoc, 'shared-bundle.html'));
-        assert.isTrue(hasImport(entrypointCDoc, 'shared-bundle.html'));
-        assert.isTrue(hasImport(shellDoc, 'shared-bundle.html'));
-      }));
-
+    // entrypoint and fragments import shared-bundle
+    assert.isTrue(hasImport(entrypointDoc, 'shared-bundle.html'));
+    assert.isTrue(hasImport(entrypointBDoc, 'shared-bundle.html'));
+    assert.isTrue(hasImport(entrypointCDoc, 'shared-bundle.html'));
+    assert.isTrue(hasImport(shellDoc, 'shared-bundle.html'));
+  }));
 });
-
-const F = (filename, contents) => new File({
-  cwd: root,
-  base: root,
-  path: path.resolve(root, filename),
-  contents: new Buffer(contents),
-});
-
-const framework = () => F('framework.html', `
-<div id="framework"></div>
-`);
-
-const shell = () => F('shell.html', `
-<link rel="import" href="framework.html">
-<div id="shell"></div>
-`);
-
-const entrypointOnly = () => F('entrypointOnly.html', `
-<link rel="import" href="/framework.html">
-`);
-
-const entrypointA = () => F('entrypointA.html', `
-<link rel="import" href="framework.html">
-<div id="entrypointA"></div>
-`);
-
-const entrypointB = () => F('entrypointB.html', `
-<link rel="import" href="commonDep.html">
-<div id="entrypointB"></div>
-`);
-
-const entrypointC = () => F('entrypointC.html', `
-<link rel="import" href="commonDep.html">
-<div id="entrypointC"></div>
-`);
-
-const commonDep = () => F('commonDep.html', `
-<div id="commonDep"></div>
-`);